Enable Text User Interface
Enable TUI so it's possible to debug using different layouts
(source/assembly etc.)
This change also uses internal version of readline (ver. 6.0) instead of the one
installed on the system(5.2). This increases size of package by a few hundreds
kilobytes. If readline version is upgraded o 6.0+ it will be possible to use
system readline again (--with-system-readline configure option)

Change-Id: If7327aac8fed9ea05624bb4b56cecf5230f9c929
Signed-off-by: Jacek Bukarewicz <j.bukarewicz@samsung.com>